web-animations/animation-node/animation-node-timing.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation node timing attribute test</title>
<meta name="assert" content="Returns the input timing properties specified for this animation node">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationNode-timing">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var node = new Animation(createDiv(this), []);

    assert_equals(node.timing.delay, 0, 'Default value of AnimationNode.timing.delay should be 0');
    assert_equals(node.timing.endDelay, 0, 'Default value of AnimationNode.timing.endDelay ' +
        'should be 0');
    assert_equals(node.timing.fill, 'auto', 'Default value of AnimationNode.timing.fill ' +
        'should be \'auto\'');
    assert_equals(node.timing.iterationStart, 0, 'Default value of ' +
        'AnimationNode.timing.iterationStart should be 0.0');
    assert_equals(node.timing.iterations, 1, 'Default value of ' +
        'AnimationNode.timing.iterations should be 1.0');
    assert_equals(node.timing.duration, 'auto', 'Default value of AnimationNode.timing.duration ' +
        'should be \'auto\'');
    assert_equals(node.timing.playbackRate, 1, 'Default value of ' +
        'AnimationNode.timing.playbackRate should be 1.0');
    assert_equals(node.timing.direction, 'normal', 'Default value of AnimationNode.timing.direction ' +
        'should be \'normal\'');
    assert_equals(node.timing.easing, 'linear', 'Default value of AnimationNode.timing.easing ' +
        'should be \'linear\'');
}, 'Test AnimationNode.timing default values');

test(function() {
    var timing = {delay: 1,
        endDelay: 1,
        fill: 'none',
        iterationStart: 1.1,
        iterations: 2.0,
        duration: 3,
        playbackRate: -2,
        playbackDirection: 'reverse',
        direction: 'reverse',
        easing: 'ease'};
    var node = new Animation(createDiv(this), [], timing);

    assert_equals(node.timing.delay, timing.delay, 'Value of AnimationNode.timing.delay ' +
        'should be specified by AnimationTimingInput');
    assert_equals(node.timing.endDelay, timing.endDelay, 'Value of AnimationNode.timing.endDelay ' +
        'should be specified by AnimationTimingInput');
    assert_equals(node.timing.fill, timing.fill, 'Value of AnimationNode.timing.fill ' +
        'should be specified by AnimationTimingInput');
    assert_equals(node.timing.iterationStart, timing.iterationStart, 'Value of ' +
        'AnimationNode.timing.iterationStart should be specified by AnimationTimingInput');
    assert_equals(node.timing.iterations, timing.iterations, 'Value of ' +
        'AnimationNode.timing.iterations should be specified by AnimationTimingInput');
    assert_equals(node.timing.duration, timing.duration, 'Value of AnimationNode.timing.duration ' +
        'should be specified by AnimationTimingInput');
    assert_equals(node.timing.playbackRate, timing.playbackRate, 'Value of ' +
        'AnimationNode.timing.playbackRate should be specified by AnimationTimingInput');
    assert_equals(node.timing.direction, timing.direction, 'Value of ' +
        'AnimationNode.timing.direction should be specified by AnimationTimingInput');
    assert_equals(node.timing.easing, timing.easing, 'Value of AnimationNode.timing.easing ' +
        'should be specified by AnimationTimingInput');
}, 'AnimationNode.timing returns the input timing properties specified for this animation node');

// TODO test all supported fill and easing values
// TODO test for AnimationGroup which implements AnimationNode interface as well
</script>
</body>
